+++ src/vpeol.rs
@@ -37,8 +37,8 @@ pub mod prelude {
     #[cfg(feature = "vpeol_3d")]
     pub use crate::vpeol_3d::{
         Vpeol3dCameraControl, Vpeol3dCameraMode, Vpeol3dPluginForEditor, Vpeol3dPluginForGame,
-        Vpeol3dPosition, Vpeol3dRotation, Vpeol3dScale, Vpeol3dTranslationGizmoConfig,
-        Vpeol3dTranslationGizmoMode, YoleckCameraChoices,
+        Vpeol3dPosition, Vpeol3dRotation, Vpeol3dScale, Vpeol3dSnapToPlane,
+        Vpeol3dTranslationGizmoConfig, Vpeol3dTranslationGizmoMode, YoleckCameraChoices,
     };
 }